requirements.txt
fastapi
pydantic
uvicorn[standard]
sqlmodel
email-validator
psycopg2-binary
python-multipart